requirements.txt
airtable==0.3.1
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
commonmark==0.9.1
cryptography==2.8
defusedxml==0.6.0
dictdiffer==0.8.1
Flask==1.1.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.1
jira==2.0.0
MarkupSafe==1.1.1
more-itertools==8.1.0
notion==0.0.25
numpy==1.18.1
oauthlib==3.1.0
packaging==20.1
pandas==1.0.1
pathspec==0.7.0
pbr==5.4.4
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.4
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
regex==2020.1.8
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
soupsieve==2.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==2.1.0
